Reset non-processing options after run
Options not part of the processing element retain the same
value after run, so this resets them to what they were before,
similar to the x/y plot script.

Currently I'm not resetting the model since that takes a while,
but perhaps I should? Might update this later if that's the case.